Move "sources" actions, reducers and selectors.
