+++ src/actions/index.js
@@ -10,7 +10,7 @@ import * as profileView from './profile-view';
 import * as publish from './publish';
 import * as receiveProfile from './receive-profile';
 import * as zippedProfiles from './zipped-profiles';
-import * as sources from './sources';
+import * as code from './code';
 
 export default Object.assign(
   {},
@@ -21,5 +21,5 @@ export default Object.assign(
   publish,
   receiveProfile,
   zippedProfiles,
-  sources
+  code
 );
