+++ src/selectors/code.js
@@ -7,10 +7,13 @@ import { createSelector } from 'reselect';
 import type { SourceCodeStatus, Selector } from 'firefox-profiler/types';
 import { getSourceViewFile } from './url-state';
 
-export const getSources: Selector<Map<string, SourceCodeStatus>> = (state) =>
-  state.sources;
+export const getSourceCodeCache: Selector<Map<string, SourceCodeStatus>> = (
+  state
+) => state.code.sourceCodeCache;
 
 export const getSourceViewCode: Selector<SourceCodeStatus | void> =
-  createSelector(getSources, getSourceViewFile, (sources, file) =>
-    file ? sources.get(file) : undefined
+  createSelector(
+    getSourceCodeCache,
+    getSourceViewFile,
+    (sourceCodeCache, file) => (file ? sourceCodeCache.get(file) : undefined)
   );
