+++ src/selectors/index.js
@@ -12,6 +12,7 @@ export * from './icons';
 export * from './publish';
 export * from './zipped-profiles';
 export * from './cpu';
+export * from './code';
 
 import * as app from './app';
 import { selectedThreadSelectors as selectedThread } from './per-thread';
@@ -22,6 +23,7 @@ import * as publish from './publish';
 import * as zippedProfiles from './zipped-profiles';
 import * as l10n from './l10n';
 import * as cpu from './cpu';
+import * as code from './code';
 
 const _selectorsForConsole = {
   app,
@@ -33,6 +35,7 @@ const _selectorsForConsole = {
   selectedThread,
   l10n,
   cpu,
+  code,
 };
 
 // Exports require explicit typing. Deduce the type with typeof.
